refactor(flow): keep the versioning work inside the quality thresholds
Running phpmd on the CHANGED FILES reported clean while running it over `lib`
— which is what CI does — reported three violations. A per-file invocation is
not the gate; the directory one is.

- `changeSchema()` had grown to 189 lines. Split into definitionStore(),
  versionStore() and pinColumns(). The three calls are assigned to separate
  variables before being OR'd, deliberately: written `$changed || $this->x()`
  the short-circuit would SKIP later steps as soon as one reported a change —
  creating the definition store and silently never adding the columns that
  point at it.
- `FlowVersionService` had a coupling of 13. It no longer takes IUserSession;
  the controller names the publisher from the FlowAccess it already holds,
  which avoids a second session dependency there for one string.
- `FlowRunService` passed 1000 lines again. `buildRun()` and `newUuid()` — both
  added by this change — move to `FlowRunRow`, which performs no checks at all:
  every guard runs in `queue()` before it is reached, so a check there would be
  a second copy of a rule that already exists, and the two would drift.

phpcs, phpmd (both passes), psalm and phpstan now all exit 0 in their CI form.
730 flow unit tests green.
