+++ lib/Service/Flow/FlowVersionService.php
@@ -41,7 +41,6 @@ use OCA\OpenRegister\Db\FlowMapper;
 use OCA\OpenRegister\Db\FlowVersion;
 use OCA\OpenRegister\Db\FlowVersionMapper;
 use OCP\IDBConnection;
-use OCP\IUserSession;
 use Psr\Log\LoggerInterface;
 use Throwable;
 
@@ -61,7 +60,6 @@ class FlowVersionService {
 	 * @param FlowTriggerIndex   $triggers The derived trigger rows.
 	 * @param IDBConnection      $db       Wraps each transition in one transaction.
 	 * @param LoggerInterface    $logger   Diagnostics.
-	 * @param IUserSession       $session  Names who published a version.
 	 */
 	public function __construct(
 		private readonly FlowVersionMapper $versions,
@@ -71,7 +69,6 @@ class FlowVersionService {
 		private readonly FlowTriggerIndex $triggers,
 		private readonly IDBConnection $db,
 		private readonly LoggerInterface $logger,
-		private readonly IUserSession $session,
 	) {
 
 	}//end __construct()
@@ -153,10 +150,6 @@ class FlowVersionService {
 	 */
 	public function publish(Flow $flow, ?string $publishedBy = null): FlowVersion {
 		$flowId = (string)$flow->getUuid();
-		// Resolved here rather than at each caller: an install-time import has
-		// no session and passes null, a request has one and should not have to
-		// remember to pass it.
-		$publishedBy = ($publishedBy ?? $this->session->getUser()?->getUID());
 		$graph = $this->graphOf(flow: $flow);
 
 		// Guards BEFORE the transaction opens. A refusal is not a rollback —
